+++ lib/manageiq/content/engine.rb
@@ -5,7 +5,7 @@ module ManageIQ
     class Engine < ::Rails::Engine
       isolate_namespace ManageIQ::Content
 
-      def vmdb_plugin?
+      def self.vmdb_plugin?
         true
       end